+++ packages/atomic/src/utils/initialization-utils.spec.ts
@@ -105,9 +105,7 @@ describe('InitializeBindings decorator', () => {
     it(`when "engine" is defined
     should render the content `, () => {
       component['bindings'] = {
-        engine: TestUtils.buildMockSearchAppEngine({
-          state: TestUtils.createMockState(),
-        }),
+        engine: TestUtils.buildMockSearchEngine(TestUtils.createMockState()),
         i18n: i18next,
         store: createAtomicStore(),
         interfaceElement: document.createElement('atomic-search-interface'),
@@ -127,9 +125,7 @@ describe('BindStateToController decorator', () => {
     console.error = jest.fn();
     component = {
       bindings: {
-        engine: TestUtils.buildMockSearchAppEngine({
-          state: TestUtils.createMockState(),
-        }),
+        engine: TestUtils.buildMockSearchEngine(TestUtils.createMockState()),
         i18n: i18next,
         store: createAtomicStore(),
         interfaceElement: document.createElement('atomic-search-interface'),
@@ -193,7 +189,7 @@ describe('BindStateToController decorator', () => {
     });
 
     it('should subscribe to the controller', () => {
-      spyOn(controller, 'subscribe');
+      jest.spyOn(controller, 'subscribe');
       BindStateToController('controller')(component, 'controllerState');
       component.initialize!();